chore: create jar
